+++ R/run.R
@@ -10,6 +10,7 @@
 #' @details
 #'  \tabular{ll}{
 #'   **Application name** \tab  **Keyword** \cr
+#'   Compositional Data Analysis \tab `coda` \cr
 #'   Matrix Seriation \tab `seriation` \cr
 #'  }
 #' @examples
@@ -20,7 +21,8 @@
 #' @family shiny apps
 #' @author N. Frerebeau
 #' @export
-run_app <- function(app = c("seriation"), options = list(launch.browser = interactive())) {
+run_app <- function(app = c("seriation", "coda"),
+                    options = list(launch.browser = interactive())) {
   app <- match.arg(app, several.ok = FALSE)
   appDir <- system.file(app, package = "janus")