+++ NEWS.md
@@ -2,6 +2,8 @@
 ## Internals
 * Import **arkhe** (>= 1.4.0).
 * Import **dimensio** (>= 0.5.0).
+* Import **isopleuros** (>= 1.0.0).
 * Import **kairos** (>= 2.0.2).
 * Import **khroma** (>= 1.11.0).
+* Import **nexus** (>= 0.1.0).
 * Import **tabula** (>= 3.0.1).